scripts/requirements.txt
psycopg2-binary
pyyaml==6.0
aiohttp==3.9
